Clean up provider file a wee bit
